requirements.txt
BeautifulSoup4
PyEnchant
dublib
Pillow